--- src/test/java/org/orderofthebee/workflow/ListActiveWorkflowsGetTest.java
@@ -1,19 +1,0 @@
-package org.orderofthebee.workflow;
-//
-//public class ListActiveWorkflowsGetTest {
-//
-//	private TestWebScriptServer server;
-//
-//	@Before
-//	public void setup(){
-//		server = TestWebScriptServer.getTestServer();
-//	}
-//	
-//	@Test
-//	public void testList(){
-//		// just see if it runs without error and gives a list of any size
-//		
-//	}
-//	
-//	
-//}
