+++ src/test/java/org/orderofthebee/workflow/TestListWorkflowDefinitionsGet.java
@@ -13,6 +13,7 @@ import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.orderofthebee.workflow.meta.TestMeta;
 
 public class TestListWorkflowDefinitionsGet {
