+++ src/test/java/org/orderofthebee/workflow/meta/TestMeta.java
@@ -1,4 +1,4 @@
-package org.orderofthebee.workflow;
+package org.orderofthebee.workflow.meta;
 
 public class TestMeta {
